# api-docs.yml
...
swagger: "2.0"
info:
  description: "SendIT is a courier service that helps users deliver parcels to different destinations. SendIT provides courier quotes based on weight categories."
  version: "1.0.0"
  title: "Send-IT"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "niomwungeri.fabrice@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "https://andela-dev-challenge.herokuapp.com"
basePath: "/api/v1"
...
schemes:
- "https"
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  JWT_Token:
    type: apiKey
    in: header
    name: x-access-token
paths:
...
  /users:
    get:
      summary: "Get all users"
      security:
        - JWT_Token: []
      tags:
        - "users"
        - "admin"
      operationId: allUsers
      parameters: 
        - name: x-access-token
          in: header
          required: true
          type: string
      responses:
        '200':
          description: return list of users
          schema:
            type: object
            properties:
              message:
                type: string
                example: Success
              status:
                type: number
                example: 200
  /users/{userId}:
    get:
      summary: "Fetch basic user's information"
      security:
        - JWT_Token: []
      tags:
        - "users"
      operationId: userById
      parameters: 
        - name: userId
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: return the basic user information
          schema:
            type: object
            properties:
              success:
                type: boolean
              data:
                $ref: "#/definitions/User"
...
  /users/{userId}/update:
    put:
      summary: "Update the user role of a user from admin to normal or the other way around"
      security:
        - JWT_Token: []
      tags:
        - "users"
      operationId: updateUser
      parameters: 
        - name: userId
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: return the basic user information
          schema:
            type: object
            properties:
              success:
                type: boolean
              data:
                $ref: "#/definitions/User"
  /users/{userId}/delete:
    delete:
      summary: "Update the user role of a user from admin to normal or the other way around"
      security:
        - JWT_Token: []
      tags:
        - "users"
      operationId: updateUser
      parameters: 
        - name: userId
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: return the basic user information
          schema:
            type: object
            properties:
              success:
                type: boolean
              data:
                $ref: "#/definitions/User"
...
definitions:
  User:
    type: object
    properties:
      id:
        type: string
        format: uuid
      email:
        type: string
        format: email
        example: niomwungeri@email.com
      first_name:
        type: string
      last_name:
        type: string
      username:
        type: string
      created_at:
        type: string
      updated_at:
        type: string
    required: 
      - id
      - email
      - first_name
      - last_name
      - username
      - created_at
      - updated_at
...